New: tests for new fields, updates for manager changes
