+++ test/tasks/EpicTest.java
@@ -28,31 +28,4 @@ class EpicTest {
         epic1.addSubTask(subTask);
         assertEquals(0, epic1.getSubTasks().size(), "эпик добавился в качестве подзадачи");
     }
-
-    @Test
-    void epicStatusTest() {
-        TaskManager manager = Managers.getDefault();
-        Epic epic = new Epic("epic1", "epic1_desc");
-        manager.addEpic(epic);
-
-        assertEquals(TaskStatus.NEW, epic.getStatus(), "статус после создания должен быть NEW");
-
-        SubTask subTask = new SubTask("sub1", "sub1_desc", epic);
-        subTask.setStatus(TaskStatus.IN_PROGRESS);
-        manager.addSubTask(subTask);
-
-        assertEquals(TaskStatus.IN_PROGRESS, manager.getEpicStatus(epic), "статус не изменился");
-
-        SubTask subTask1 = new SubTask("sub2", "sub2_desc", epic);
-        subTask1.setStatus(TaskStatus.DONE);
-        subTask.setStatus(TaskStatus.DONE);
-        manager.addSubTask(subTask1);
-        manager.updateSubTask(subTask);
-
-        assertEquals(TaskStatus.DONE, manager.getEpicStatus(epic), "статус не изменился");
-
-        manager.getEpic(epic.getId()).setStatus(TaskStatus.NEW);
-
-        assertNotEquals(TaskStatus.NEW, manager.getEpicStatus(epic), "статус изменился после setStatus()");
-    }
 }
